Apply useful stuff from pr
