+++ src/extension/companion/actionTemplate.ts
@@ -1,0 +1,5 @@
+// This is a template of witch you can base your own actions on.
+
+export default function actionTemplate(name: string, value: unknown): void {
+  // TODO: write code
+}
